Remove code from `pex/resolve/lockfile/__init__.py`.
The primary refactors:
+ pex.resolve.lockfile.{lockfile -> model}
+ pex.resolve.lockfile.{create -> operations.create}
+ pex.resolve.lockfile.{load,loads} inlined json_codec.{load,loads}

This is part of a series of a refactors needed to remove dependency
cycles in support of #1696.
