+++ pex/resolve/lockfile/model.py
