+++ pex/resolve/lockfile/updater.py
@@ -14,7 +14,8 @@ from pex.network_configuration import NetworkConfiguration
 from pex.pep_440 import Version
 from pex.pep_503 import ProjectName
 from pex.resolve.locked_resolve import LockConfiguration, LockedRequirement, LockedResolve
-from pex.resolve.lockfile import Lockfile, create
+from pex.resolve.lockfile.model import Lockfile
+from pex.resolve.lockfile.operations import create
 from pex.resolve.requirement_configuration import RequirementConfiguration
 from pex.resolve.resolver_configuration import PipConfiguration, ReposConfiguration
 from pex.result import Error, ResultError, catch, try_
