+++ tests/integration/cli/commands/test_lock.py
@@ -1,6 +1,7 @@
 # Copyright 2021 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
+import json
 import os
 import re
 import subprocess
@@ -10,14 +11,15 @@ from textwrap import dedent
 import pytest
 
 from pex.cli.testing import run_pex3
+from pex.common import safe_open
 from pex.dist_metadata import Requirement
 from pex.interpreter import PythonInterpreter
 from pex.pep_440 import Version
 from pex.pep_503 import ProjectName
-from pex.resolve import lockfile
 from pex.resolve.locked_resolve import Artifact, LockedRequirement
-from pex.resolve.lockfile import Lockfile
+from pex.resolve.lockfile import json_codec
 from pex.resolve.lockfile.download_manager import DownloadedArtifact
+from pex.resolve.lockfile.model import Lockfile
 from pex.resolve.resolved_requirement import Fingerprint, Pin
 from pex.resolve.resolver_configuration import ResolverVersion
 from pex.resolve.testing import normalize_locked_resolve
@@ -78,8 +80,8 @@ def test_create(tmpdir):
     # comment line which is sensitive to the source of the requirements)
     result = run_pex3("lock", "create", "-r", requirements_file)
     result.assert_success()
-    assert normalize_lockfile(lockfile.load(lock_file)) == normalize_lockfile(
-        lockfile.loads(result.output)
+    assert normalize_lockfile(json_codec.load(lock_file)) == normalize_lockfile(
+        json_codec.loads(result.output)
     )
 
 
@@ -109,7 +111,7 @@ def test_create_style(
             pex_root,
         )
         run_pex3(*(args + additional_args)).assert_success()
-        lock = lockfile.load(lock_file)
+        lock = json_codec.load(lock_file)
         assert 1 == len(lock.locked_resolves)
         locked_resolve = lock.locked_resolves[0]
         assert 1 == len(locked_resolve.locked_requirements)
@@ -363,7 +365,7 @@ UPDATE_LOCKFILE_CONTENTS = """\
 """
 
 
-UPDATE_LOCKFILE = lockfile.loads(UPDATE_LOCKFILE_CONTENTS)
+UPDATE_LOCKFILE = json_codec.loads(UPDATE_LOCKFILE_CONTENTS)
 
 
 def write_lock_file(
@@ -420,7 +422,7 @@ def test_update_noop(lock_file_path):
         "There were no updates for urllib3 in lock generated by cp38-cp38-manylinux_2_33_x86_64.\n"
         == result.error
     )
-    assert UPDATE_LOCKFILE == lockfile.load(lock_file_path)
+    assert UPDATE_LOCKFILE == json_codec.load(lock_file_path)
 
 
 def test_update_noop_dry_run(lock_file_path):
@@ -448,7 +450,7 @@ def test_update_targeted_upgrade_miss(lock_file_path):
 
 def test_update_targeted_upgrade(lock_file_path):
     # type: (str) -> None
-    assert SortedTuple() == lockfile.load(lock_file_path).constraints
+    assert SortedTuple() == json_codec.load(lock_file_path).constraints
     result = run_lock_update_for_py310("-p", "urllib3<1.26.7", lock_file_path)
     result.assert_success()
     assert not result.output
@@ -457,7 +459,7 @@ def test_update_targeted_upgrade(lock_file_path):
         "cp38-cp38-manylinux_2_33_x86_64.\n" == result.error
     )
 
-    lock_file = lockfile.load(lock_file_path)
+    lock_file = json_codec.load(lock_file_path)
     assert SortedTuple([Requirement.parse("urllib3<1.26.7")]) == lock_file.constraints
     assert 1 == len(lock_file.locked_resolves)
     locked_resolve = lock_file.locked_resolves[0]
@@ -483,7 +485,7 @@ def test_update_targeted_upgrade_dry_run(lock_file_path):
         "cp38-cp38-manylinux_2_33_x86_64.\n" == result.output
     )
     assert not result.error
-    assert UPDATE_LOCKFILE == lockfile.load(
+    assert UPDATE_LOCKFILE == json_codec.load(
         lock_file_path
     ), "A dry run update should not have updated the lock file."
 
@@ -508,7 +510,7 @@ def test_update_targeted_closure_shrink(lock_file_path):
 
     result = run_lock_update_for_py310("-p", "requests==2.0.0", lock_file_path)
     result.assert_success()
-    lock_file = lockfile.load(lock_file_path)
+    lock_file = json_codec.load(lock_file_path)
     assert 1 == len(lock_file.locked_resolves)
     locked_resolve = lock_file.locked_resolves[0]
     assert [
@@ -557,10 +559,14 @@ def test_update_targeted_impossible(
 
     # The pip legacy resolver, though is not strict and will let us get away with this.
     updated_lock_file_path = os.path.join(str(tmpdir), "lock.updated")
-    lockfile.store(
-        attr.evolve(UPDATE_LOCKFILE, resolver_version=ResolverVersion.PIP_LEGACY),
-        updated_lock_file_path,
-    )
+    with safe_open(updated_lock_file_path, "w") as fp:
+        json.dump(
+            json_codec.as_json_data(
+                attr.evolve(UPDATE_LOCKFILE, resolver_version=ResolverVersion.PIP_LEGACY)
+            ),
+            fp,
+            sort_keys=True,
+        )
     result = run_lock_update_for_py310("-p", "urllib3<1.16", updated_lock_file_path)
     result.assert_success()
     assert not result.output
@@ -637,7 +643,7 @@ DUAL_UPDATE_LOCKFILE_CONTENTS = """\
 """
 
 
-DUAL_UPDATE_LOCKFILE = lockfile.loads(DUAL_UPDATE_LOCKFILE_CONTENTS)
+DUAL_UPDATE_LOCKFILE = json_codec.loads(DUAL_UPDATE_LOCKFILE_CONTENTS)
 
 
 def test_update_partial(tmpdir):
@@ -653,7 +659,7 @@ def test_update_partial(tmpdir):
         lock_file_path,
     )
     result.assert_success()
-    assert DUAL_UPDATE_LOCKFILE == lockfile.load(lock_file_path)
+    assert DUAL_UPDATE_LOCKFILE == json_codec.load(lock_file_path)
 
     # By default, lock updates are strict: all locked resolves must be updated at once.
     result = run_lock_update(
@@ -679,7 +685,7 @@ def test_update_partial(tmpdir):
         lock_file_path,
     )
     result.assert_success()
-    assert DUAL_UPDATE_LOCKFILE == lockfile.load(lock_file_path)
+    assert DUAL_UPDATE_LOCKFILE == json_codec.load(lock_file_path)
 
 
 def test_excludes_pep517_build_requirements_issue_1565(tmpdir):
@@ -699,7 +705,7 @@ def test_excludes_pep517_build_requirements_issue_1565(tmpdir):
         "--force-pep517",
     )
     result.assert_success()
-    lock = lockfile.loads(result.output)
+    lock = json_codec.loads(result.output)
 
     assert 1 == len(lock.locked_resolves)
     assert (
@@ -1394,8 +1400,8 @@ PIP_2020_UNIVERSAL_LOCKFILE_CONTENTS = """\
 # N.B.: These two locks have the same contents but differing artifact order. The Pip legacy resolver
 # sorts cffi and cryptography mac wheels before win, and the 2020 resolver does the reverse.
 EXPECTED_LOCKFILES = {
-    ResolverVersion.PIP_LEGACY: lockfile.loads(LEGACY_UNIVERSAL_LOCKFILE_CONTENTS),
-    ResolverVersion.PIP_2020: lockfile.loads(PIP_2020_UNIVERSAL_LOCKFILE_CONTENTS),
+    ResolverVersion.PIP_LEGACY: json_codec.loads(LEGACY_UNIVERSAL_LOCKFILE_CONTENTS),
+    ResolverVersion.PIP_2020: json_codec.loads(PIP_2020_UNIVERSAL_LOCKFILE_CONTENTS),
 }
 
 
@@ -1455,7 +1461,7 @@ def test_universal_lock(
         cwd=os.path.dirname(constraints_file),
     )
     result.assert_success()
-    lock = lockfile.loads(result.output)
+    lock = json_codec.loads(result.output)
 
     platform_tag = py39.identity.supported_tags[0]
     assert (
